Update 9th Gate for Sinar Pure Foods International Preparation for Go Live Finished Goods Modules 6th Commit for Finished Goods Modules
Updates:
+ Non-Fish
	+ Stores Withdrawal
		/ stock on hand field (soh) from int to numeric (18, 2)
		
+ Finished Goods
	+ Pallet Transfer
		+ add support on hold state
		+ add reject state
		/ quantity shown on list using cs and tin
		/ quantity on create, separate between cs and tin
	+ Order Fill
		+ validate on same PTS Number, BOR yet picked, empty PTS Number
		+ show flakes information
		+ special instruction (same with remarks)
		+ validate maximum can code can entered
		/ quantity on create, separate between cs and tin
	+ Labeling Monitoring
		+ bad stocks category: flipper, bulger and false seam
		+ other category: no code, blurred code and liquor mentah
		/ quantity on create, separate between cs and tin
	+ Export Delivery
		/ label pelayaran seal to seal number
	+ Transfer
		/ list item re-production to re-process
		/ lise department from/to change to input free text
	+ Reclassification
		/ item reclassification target limit to pack_size only
		/ quantity on create, separate between cs and tin
	+ Delivery
		/ quantity on create, separate between cs and tin
	+ Pallet Disposition
		+ create function
		+ list function
	+ Pallet Rejection
		+ create function
		+ list function
	+ Item
		+ item oil and item lid data
		
+ Report
	+ Finished Goods
		+ Stock Inventory
			+ preview data function
			/ query and view to get OUT_DELIVERY

+++ src/java/com/app/wms/engine/db/dao/spring/SwsDtlDaoImpl.java
@@ -32,7 +32,7 @@ public class SwsDtlDaoImpl extends AbstractDAO
         sd.setSwsCode(rs.getString("sws_code"));
         sd.setProductCode(rs.getString("product_code"));
         sd.setQty(rs.getBigDecimal("qty"));
-        sd.setSoh(rs.getInt("soh"));
+        sd.setSoh(rs.getBigDecimal("soh"));
         sd.setUom(rs.getString("uom"));
         sd.setCreatedBy(rs.getString("created_by"));
         sd.setCreatedDate(rs.getDate("created_date"));
